refactor(components)
* refactor components code
* remove unused Router.js
* fix serviceWorker.js with Eslint rules
