+++ src/components/FavoritesPostsContainer/index.js
@@ -1,0 +1,1 @@
+export { FavoritesPostsContainer } from './FavoritesPostsContainer';
