+++ src/components/Header/Header.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+
 import './header.css';
 
 export const Header = ({ children }) => (
