--- src/components/Loader/Loader.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import './loader.css';
-
-export const Loader = () => {
-  return (
-    <div className="lds-default">
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-      <div/>
-    </div>
-  )
-};
