+++ src/components/Loader/Loader.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import './loader.css';
 
 export const Loader = () => (
