+++ src/components/Post/index.js
@@ -1,0 +1,1 @@
+export { Post } from './Post';
