+++ src/components/PostsContainer/index.js
@@ -1,0 +1,1 @@
+export { PostsContainer } from './PostsContainer';
