+++ src/components/index.js
@@ -1,5 +1,10 @@
 export { CustomButton } from './CustomButton';
+export { FavoritesPostsContainer } from './FavoritesPostsContainer';
 export { Header } from './Header';
+export { Loader } from './Loader';
 export { Menu } from './Menu';
+export { Modal } from './Modal';
+export { Post } from './Post';
+export { PostsContainer } from './PostsContainer';
 export { Search } from './Search';
 export { SelectedPost } from './SelectedPost';
